[FEATURE] #17 - 경로 검색 지도에 반영

+++ core/data/src/main/java/com/eight_potato/data/model/address/PoiData.kt
@@ -3,8 +3,8 @@ package com.eight_potato.data.model.address
 import com.eight_potato.domain.model.address.PoiModel
 
 data class PoiData(
-    val lat: Double, // 위도
-    val lon: Double // 경도
+    val lat: Float, // 위도
+    val lon: Float // 경도
 ) {
     override fun toString(): String {
         return "$lat,$lon"
